refactor(backend): move oauth2 usecases to right place
